+++ Original/notifica/src/main/java/br/com/bd_notifica/repositories/CriacaoDeTicketRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.bd_notifica.entities.Ticket;
 
-public interface CriacaoDeTicketRepository extends JpaRepository<Ticket, Long> {
+public interface CriacaoDeTicketRepository extends JpaRepository<Ticket, Integer> {
 
 }
